# src/department/dto/create-department.dto.ts
import { ApiProperty, ApiPropertyOptional } from '@nestjs/swagger';
import { IsString, IsOptional, MinLength } from 'class-validator';

export class CreateDepartmentDto {
  @ApiProperty({ description: 'Organization this department belongs to' })
  @IsString()
  organizationId: string;

  @ApiProperty({ example: 'Engineering' })
  @IsString()
  @MinLength(1)
  name: string;

  @ApiPropertyOptional({ example: 'Builds and maintains the product' })
  @IsOptional()
  @IsString()
  description?: string;

  @ApiPropertyOptional({
    description:
      'Employee ID of the department manager. Must belong to the same organization and cannot already manage another department.',
  })
  @IsOptional()
  @IsString()
  managerId?: string;
}